<fix>(Filter preselection): use preselection in Dropdown (needed to happen on load)
